requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
lightgbm>=4.0.0
xgboost>=2.0.0
torch>=2.0.0
chronos-forecasting>=1.0.0
streamlit>=1.28.0
plotly>=5.15.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0
pyyaml>=6.0.0
